Prepare for meson: rename assert.h & endian.h
These shadow system header file names, which is a bad
practice in general.
Unfortunately due to how meson builds include directories
they get preferred for including over the system headers.
