+++ src/common/caspar_assert.h
