+++ src/modules/newtek/producer/newtek_ndi_producer.cpp
@@ -29,7 +29,7 @@
 #include <core/monitor/monitor.h>
 #include <core/producer/frame_producer.h>
 
-#include <common/assert.h>
+#include <common/caspar_assert.h>
 #include <common/diagnostics/graph.h>
 #include <common/env.h>
 #include <common/executor.h>
